+++ src/helpers/index.js
@@ -1,4 +1,5 @@
 import normalizeErrorResponse from './normalizeErrorResponse';
 import normalizeDocument from './normalizeDocument';
+import readJsonAtRoot from './readJsonAtRoot';
 
-export { normalizeDocument, normalizeErrorResponse };
+export { normalizeDocument, normalizeErrorResponse, readJsonAtRoot };