+++ src/middlewares/index.js
@@ -1,8 +1,7 @@
-import useStorage from './useStorage';
 import useInstrumentation from './useInstrumentation';
 import useHttpLogger from './useHttpLogger';
 import traceRequest from './traceRequest';
 
 export {
-  useStorage, useInstrumentation, useHttpLogger, traceRequest,
+  useInstrumentation, useHttpLogger, traceRequest,
 };