+++ tests/ServiceClientFactory.test.js
@@ -1,28 +1,46 @@
 import nock from 'nock';
+import mockFs from 'mock-fs';
 import ServiceClientFactory from '../src/ServiceClientFactory';
 
 describe('ServiceClientFactory', () => {
+  beforeAll(() => {
+    mockFs({
+      'service-env.json': JSON.stringify({
+        id: 'order-service',
+        name: 'order service',
+        dependencies: [{
+          id: 'storage-service',
+          name: 'storage service',
+          category: 'storage',
+          endpoint: 'http://storage-service:3000',
+          version: '1.0.0',
+        }],
+      }),
+    });
+  });
+
+  afterEach(() => {
+    mockFs.restore();
+  });
+
   afterEach(() => {
     nock.restore();
   });
 
-  it('can get "storageServiceClient"', async () => {
-    nock('http://registry-service:3000')
-      .get('/services/storage-service')
+  it('can execute request to targeted service correctly', async () => {
+    nock('http://storage-service:3000')
+      .get('/healthz')
       .reply(200, {
-        data: {
-          type: 'services',
-          attributes: {
-            endpoint: 'http://storage-service:3001',
-          },
-        },
+        ping: 'pong',
       });
 
+    const storageServiceClient = await ServiceClientFactory.create('storage-service');
 
-    const storageServiceClient = new ServiceClientFactory({
-      name: 'storage-service',
+    const response = await storageServiceClient.request({
+      method: 'get',
+      url: '/healthz',
     });
 
-    expect(storageServiceClient.request).toBeInstanceOf(Function);
+    expect(response.status).toBe(200);
   });
 });